Add values, tags support, split merger into 2 files
